feat: mobile professor dropdown menu is responsive
